+++ dev.ts
@@ -5,7 +5,4 @@ import config from "./fresh.config.ts";
 
 import "$std/dotenv/load.ts";
 
-// Set build environment flag
-Deno.env.set("FRESH_BUILD", "1");
-
 await dev(import.meta.url, "./main.ts", config);
