--- islands/CandleStand.tsx
@@ -1,8 +1,0 @@
-import Candle from "../components/Candle.tsx";
-
-export default function CandleStand() {
-  return (
-    <div class="w-full h-[200px] rounded-lg candle-stand">
-    </div>
-  );
-}
