+++ routes/_layout.tsx
@@ -1,6 +1,5 @@
 import { PageProps } from "$fresh/server.ts";
 import { Link } from "../components/Link.tsx";
-import { Status } from "../islands/Status.tsx";
 
 export default function Layout({ Component }: PageProps) {
   // do something with state here
@@ -9,7 +8,6 @@ export default function Layout({ Component }: PageProps) {
       <h1 class="text-3xl/8 font-bold font-triodion">
         lofi byzantine music radio &mdash; chanting to relax/study to
         <span class="left-1 bottom-4 relative">
-          {/* <Status /> */}
         </span>
       </h1>
       <main className={`grow`}>
