+++ routes/api/radio.ts
@@ -21,8 +21,8 @@ const clients = new Set<WebSocket>();
 
 const DEV_ENV = Deno.env.get("ENVIRONMENT") === "DEV";
 export const AUDIO_DIRECTORY = DEV_ENV
-  ? "./static/audio"
-  : "/home/george/media/orthodox/albums";
+  ? "./static/audio/paraklesis-byzantine-ark"
+  : "/home/george/media/orthodox/mp3";
 
 // Track the current file being served to all clients
 let currentTrack: AudioTrack | null = null;
