+++ routes/index.tsx
@@ -1,7 +1,6 @@
 import { Head } from "$fresh/runtime.ts";
-import AudioPlayer from "../islands/AudioPlayer.tsx";
 import LatestCommit from "../islands/LatestCommit.tsx";
-import NewAudioPlayer from "../islands/PlayPauseButton.tsx";
+import AudioPlayer from "../islands/AudioPlayer.tsx";
 
 export default function Home() {
   return (
@@ -31,10 +30,6 @@ export default function Home() {
 
       <div>
         <AudioPlayer />
-      </div>
-
-      <div>
-        <NewAudioPlayer />
       </div>
     </>
   );
